fix: 동시성 테스트 통합,add ConcurrentHashMap
